* .go spawn 1 should save to db now * uncommented .go info
git-svn-id: svn://svn.aspiredev.org/AspireCore/branches/Hearthstone@578 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/hearthstone-world/GameObject.cpp
@@ -267,7 +267,7 @@ void GameObject::SaveToDB()
 		<< GetFloatValue(GAMEOBJECT_ROTATION_01) << ","
 		<< GetFloatValue(GAMEOBJECT_ROTATION_02) << ","
 		<< GetFloatValue(GAMEOBJECT_ROTATION_03) << ","
-		<< GetByte(GAMEOBJECT_BYTES_1, GAMEOBJECT_BYTES_STATE) << ","
+		<< (uint32)GetByte(GAMEOBJECT_BYTES_1, GAMEOBJECT_BYTES_STATE) << ","
 		<< GetUInt32Value(GAMEOBJECT_FLAGS) << ","
 		<< GetUInt32Value(GAMEOBJECT_FACTION) << ","
 		<< GetFloatValue(OBJECT_FIELD_SCALE_X) << ","
